Extract setup db
