+++ chronologger.go
@@ -1,32 +1,18 @@
 package main
 
 import (
-	"context"
 	"time"
 
 	"github.com/jaswdr/faker"
 	"github.com/labstack/echo/v4"
 	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
+
+	db "github.com/hay-i/chronologger/db"
 )
 
 func main() {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	clientOpts := options.Client().ApplyURI("mongodb://root:example@localhost:27017")
-	client, err := mongo.Connect(ctx, clientOpts)
-
-	defer func() {
-		if err = client.Disconnect(ctx); err != nil {
-			panic(err)
-		}
-	}()
-
-	if err != nil {
-		panic(err)
-	}
+	ctx, client := db.SetupDb()
 
 	collection := client.Database("chronologger").Collection("goals")
 	fake := faker.New()
