+++ db/setup.go
@@ -1,0 +1,29 @@
+package db
+
+import (
+	"context"
+	"time"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func SetupDb() (context.Context, *mongo.Client) {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	clientOpts := options.Client().ApplyURI("mongodb://root:example@localhost:27017")
+	client, err := mongo.Connect(ctx, clientOpts)
+
+	defer func() {
+		if err = client.Disconnect(ctx); err != nil {
+			panic(err)
+		}
+	}()
+
+	if err != nil {
+		panic(err)
+	}
+
+	return ctx, client
+}
